Update user layout file
Changed the title from the foundation default
Made a crude if statement to set the user_id or just id for the user
editing or stem creation views respectively.
Activated the offcanvas nav links to those views.
Extracted the flash success and/or error messages into this layout
instead of at the top of the 2 main forms created so far (users and
stems.)